Kryefaqja

+++ app/Http/Controllers/AutoriController.php
@@ -7,6 +7,12 @@ use App\Autor;
 
 class AutoriController extends Controller
 {
+
+    public function autoret()
+    {
+        $autoret = Autor::all();
+        return view('bibloteka.autoret')->with('autoret',$autoret);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -14,8 +20,12 @@ class AutoriController extends Controller
      */
     public function index()
     {
+        if(!auth()->user()->isAdmin)
+            return redirect('/')->with('error','Nuk keni autorizim');
         $autoret = Autor::all();
         return view('autori.index')->with('autoret',$autoret);
+
+
     }
 
     /**
@@ -25,7 +35,10 @@ class AutoriController extends Controller
      */
     public function create()
     {
+        if(!auth()->user()->isAdmin)
+            return redirect('/')->with('error','Nuk keni autorizim');
         return view('autori.create');
+
     }
 
     /**
@@ -36,7 +49,8 @@ class AutoriController extends Controller
      */
     public function store(Request $request)
     {
-
+        if(!auth()->user()->isAdmin)
+            return redirect('/')->with('error','Nuk keni autorizim');
         $request->validate([
             'name' => 'required',
             'periudha' => 'required',
@@ -56,6 +70,8 @@ class AutoriController extends Controller
      */
     public function show($id)
     {
+        if(!auth()->user()->isAdmin)
+            return redirect('/')->with('error','Nuk keni autorizim');
         $autori = Autor::find($id);
         return view('autori.show')->with('autori',$autori);
     }
@@ -68,6 +84,8 @@ class AutoriController extends Controller
      */
     public function edit($id)
     {
+        if(!auth()->user()->isAdmin)
+            return redirect('/')->with('error','Nuk keni autorizim');
         $autori = Autor::find($id);
         return view('autori.edit')->with('autori',$autori);
     }
@@ -81,6 +99,8 @@ class AutoriController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(!auth()->user()->isAdmin)
+            return redirect('/')->with('error','Nuk keni autorizim');
         $request->validate([
             'name' => 'required',
             'periudha' => 'required',
@@ -100,6 +120,8 @@ class AutoriController extends Controller
      */
     public function destroy($id)
     {
+        if(!auth()->user()->isAdmin)
+            return redirect('/')->with('error','Nuk keni autorizim');
         $autori = Autor::find($id);
         $autori->delete();
         return redirect('/autor');
